refactor: placeItem 분리, type 필드 추가
