+++ src/main/java/backend/capstone/domain/dayroute/dto/DayRouteDetailResponse.java
@@ -1,5 +1,6 @@
 package backend.capstone.domain.dayroute.dto;
 
+import backend.capstone.domain.place.dto.PlaceItem;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.Builder;
@@ -15,17 +16,4 @@ public record DayRouteDetailResponse(
     Integer pathPointCount,
     List<PlaceItem> places
 ) {
-
-    @Builder
-    public record PlaceItem(
-        Long placeId,
-        String placeName,
-        String roadAddress,
-        double latitude,
-        double longitude,
-        int orderIndex
-    ) {
-
-    }
-
 }
