+++ src/main/java/backend/capstone/domain/dayroute/mapper/DayRouteMapper.java
@@ -5,6 +5,7 @@ import backend.capstone.domain.dayroute.dto.DayRouteMonthlyResponse;
 import backend.capstone.domain.dayroute.dto.GpsPointsResponse;
 import backend.capstone.domain.dayroute.entity.DayRoute;
 import backend.capstone.domain.gpspoint.entity.GpsPoint;
+import backend.capstone.domain.place.dto.PlaceItem;
 import backend.capstone.domain.place.entity.Place;
 import backend.capstone.domain.user.entity.User;
 import java.time.LocalDate;
@@ -49,9 +50,10 @@ public class DayRouteMapper {
             .encodedPath(dayRoute.getEncodedPath())
             .pathPointCount(dayRoute.getPathPointCount())
             .places(places.stream()
-                .map(p -> DayRouteDetailResponse.PlaceItem.builder()
+                .map(p -> PlaceItem.builder()
                     .placeId(p.getId())
                     .placeName(p.getName())
+                    .type(p.getSource())
                     .roadAddress(p.getRoadAddress())
                     .latitude(p.getLatitude())
                     .longitude(p.getLongitude())
